# Improve performance of `is_dependency_of` query