### Description

<!-- REQUIRED
    Provide a concise description of your change. What does it do? Why is it necessary?
    As a guideline, think about how you would describe your change if you were to write a changelog entry for it.
-->

It is generally to be expected that the vast majority of components will *not* match the search criteria of `is_dependency_of`. Thus, it makes sense to optimize for the negative case.

The following additions were made:
* Perform a pre-flight check whether any component in the project matches the search criteria at all; The old query would traverse the entire graph even though no component could ever match
* Stop recursing once a match is found; The old query would keep note of when a component matched, but not abort the recursion, causing unnecessary queries to be performed
* Replace raw JDBC usage with JDBI :)

Queries where *no* component in the project matches the desired criteria complete in less than a millisecond.

Execution times of queries where at least one component matches the criteria now directly correlates with how "close" to the leaf node those components are.

### Addressed Issue

<!-- REQUIRED
    Reference the issue addressed by this PR, e.g. `#1234`.
    Use keywords like `closes` or `fixes` to signal that this PR resolves the issue,
    causing the issue to be closed automatically when the PR is merged:
        https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
-->

N/A

### Additional Details

<!-- OPTIONAL
    If desired, share more technical details about the change here.
    Elaborating on why you implemented the change the way you did can be super helpful to the reviewer.
    Did you consider other solutions? Any problems you ran into along the way?
-->

I did some testing, using a portfolio of 2000 projects ~700'000 components in total.
I used the [`bloated.bom.json`](https://github.com/DependencyTrack/hyades/blob/main/load-tests/fixtures/boms/bloated.bom.json) as test subject.

For testing positive matches, I selected `is-ci@2.0.0` as leaf node, and `miguelcostero-ng2-toasty@0.0.0-semantically-released` as root node. The distance between those components is 11. The query plan for both old and new query can be found below:

<details>
<summary>Old (114.858 ms)</summary>
<pre>
+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|QUERY PLAN                                                                                                                                                                               |
+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|Aggregate  (cost=38180.43..38180.44 rows=1 width=1) (actual time=114.789..114.790 rows=1 loops=1)                                                                                        |
|  CTE CTE_DEPENDENCIES                                                                                                                                                                   |
|    ->  Recursive Union  (cost=231.53..38143.71 rows=1632 width=78) (actual time=10.248..114.770 rows=11 loops=1)                                                                        |
|          ->  Bitmap Heap Scan on "COMPONENT" "C"  (cost=231.53..262.65 rows=12 width=78) (actual time=10.247..10.650 rows=1 loops=1)                                                    |
|                Recheck Cond: ("DIRECT_DEPENDENCIES" ~~ '%89dc699b-1518-49f9-ba74-04a8d0234400%'::text)                                                                                  |
|                Rows Removed by Index Recheck: 1                                                                                                                                         |
|                Heap Blocks: exact=2                                                                                                                                                     |
|                ->  Bitmap Index Scan on idx_gin  (cost=0.00..231.53 rows=28 width=0) (actual time=10.237..10.237 rows=2 loops=1)                                                        |
|                      Index Cond: ("DIRECT_DEPENDENCIES" ~~ '%89dc699b-1518-49f9-ba74-04a8d0234400%'::text)                                                                              |
|          ->  Nested Loop  (cost=29.28..3786.47 rows=162 width=78) (actual time=9.352..9.463 rows=1 loops=11)                                                                            |
|                ->  WorkTable Scan on "CTE_DEPENDENCIES" "CTE_DEPENDENCIES_1"  (cost=0.00..2.40 rows=120 width=131) (actual time=0.000..0.000 rows=1 loops=11)                           |
|                ->  Bitmap Heap Scan on "COMPONENT" "C_1"  (cost=29.28..31.52 rows=1 width=674) (actual time=9.343..9.454 rows=1 loops=11)                                               |
|                      Recheck Cond: (("CTE_DEPENDENCIES_1"."PROJECT_ID" = "PROJECT_ID") AND ("DIRECT_DEPENDENCIES" ~~ (('%'::text || ("CTE_DEPENDENCIES_1"."UUID")::text) || '%'::text)))|
|                      Rows Removed by Index Recheck: 1                                                                                                                                   |
|                      Heap Blocks: exact=18                                                                                                                                              |
|                      ->  BitmapAnd  (cost=29.28..29.28 rows=2 width=0) (actual time=9.123..9.123 rows=0 loops=11)                                                                       |
|                            ->  Bitmap Index Scan on "COMPONENT_PROJECT_ID_IDX"  (cost=0.00..4.33 rows=379 width=0) (actual time=0.212..0.212 rows=9056 loops=11)                        |
|                                  Index Cond: ("PROJECT_ID" = "CTE_DEPENDENCIES_1"."PROJECT_ID")                                                                                         |
|                            ->  Bitmap Index Scan on idx_gin  (cost=0.00..24.70 rows=3409 width=0) (actual time=8.871..8.871 rows=2 loops=11)                                            |
|                                  Index Cond: ("DIRECT_DEPENDENCIES" ~~ (('%'::text || ("CTE_DEPENDENCIES_1"."UUID")::text) || '%'::text))                                               |
|  ->  CTE Scan on "CTE_DEPENDENCIES"  (cost=0.00..32.64 rows=1632 width=1) (actual time=10.249..114.778 rows=11 loops=1)                                                                 |
|Planning Time: 0.462 ms                                                                                                                                                                  |
|Execution Time: 114.858 ms                                                                                                                                                               |
+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
</pre>
</details>

<details>
<summary>New (111.331 ms)</summary>
<pre>
+---------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|QUERY PLAN                                                                                                                                                           |
+---------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|CTE Scan on "CTE_DEPENDENCIES"  (cost=1782.44..1784.26 rows=91 width=131) (actual time=4.648..111.203 rows=11 loops=1)                                               |
|  CTE CTE_PROJECT                                                                                                                                                    |
|    ->  Index Scan using "COMPONENT_UUID_IDX" on "COMPONENT" "C"  (cost=0.42..2.64 rows=1 width=8) (actual time=0.016..0.017 rows=1 loops=1)                         |
|          Index Cond: (("UUID")::text = '89dc699b-1518-49f9-ba74-04a8d0234400'::text)                                                                                |
|  CTE CTE_MATCHES                                                                                                                                                    |
|    ->  Bitmap Heap Scan on "COMPONENT"  (cost=6.61..7.73 rows=1 width=8) (actual time=0.585..0.586 rows=1 loops=1)                                                  |
|          Recheck Cond: ((("NAME")::text = 'miguelcostero-ng2-toasty'::text) AND ("PROJECT_ID" = $1))                                                                |
|          Filter: ("DIRECT_DEPENDENCIES" IS NOT NULL)                                                                                                                |
|          Heap Blocks: exact=1                                                                                                                                       |
|          InitPlan 2 (returns $1)                                                                                                                                    |
|            ->  CTE Scan on "CTE_PROJECT"  (cost=0.00..0.02 rows=1 width=8) (actual time=0.018..0.018 rows=1 loops=1)                                                |
|          ->  BitmapAnd  (cost=6.59..6.59 rows=1 width=0) (actual time=0.579..0.580 rows=0 loops=1)                                                                  |
|                ->  Bitmap Index Scan on "COMPONENT_NAME_IDX"  (cost=0.00..1.98 rows=60 width=0) (actual time=0.021..0.021 rows=1 loops=1)                           |
|                      Index Cond: (("NAME")::text = 'miguelcostero-ng2-toasty'::text)                                                                                |
|                ->  Bitmap Index Scan on "COMPONENT_PROJECT_ID_IDX"  (cost=0.00..4.37 rows=379 width=0) (actual time=0.557..0.557 rows=9056 loops=1)                 |
|                      Index Cond: ("PROJECT_ID" = $1)                                                                                                                |
|  CTE CTE_DEPENDENCIES                                                                                                                                               |
|    ->  Recursive Union  (cost=0.49..1772.07 rows=91 width=78) (actual time=4.646..111.191 rows=11 loops=1)                                                          |
|          ->  Result  (cost=0.49..174.76 rows=1 width=78) (actual time=4.645..12.423 rows=1 loops=1)                                                                 |
|                One-Time Filter: $5                                                                                                                                  |
|                InitPlan 5 (returns $5)                                                                                                                              |
|                  ->  CTE Scan on "CTE_MATCHES" "CTE_MATCHES_1"  (cost=0.00..0.02 rows=1 width=0) (actual time=0.586..0.586 rows=1 loops=1)                          |
|                InitPlan 6 (returns $6)                                                                                                                              |
|                  ->  CTE Scan on "CTE_PROJECT" "CTE_PROJECT_1"  (cost=0.00..0.02 rows=1 width=8) (actual time=0.000..0.000 rows=1 loops=1)                          |
|                ->  Index Scan using "COMPONENT_PROJECT_ID_IDX" on "COMPONENT" "C_1"  (cost=0.42..174.70 rows=1 width=53) (actual time=4.043..11.819 rows=1 loops=1) |
|                      Index Cond: ("PROJECT_ID" = $6)                                                                                                                |
|                      Filter: ("DIRECT_DEPENDENCIES" ~~ '%89dc699b-1518-49f9-ba74-04a8d0234400%'::text)                                                              |
|                      Rows Removed by Filter: 9055                                                                                                                   |
|                SubPlan 4                                                                                                                                            |
|                  ->  CTE Scan on "CTE_MATCHES"  (cost=0.00..0.02 rows=1 width=8) (actual time=0.001..0.002 rows=1 loops=1)                                          |
|          ->  Nested Loop  (cost=29.61..159.64 rows=9 width=78) (actual time=8.875..8.976 rows=1 loops=11)                                                           |
|                ->  WorkTable Scan on "CTE_DEPENDENCIES" "PREVIOUS"  (cost=0.00..0.20 rows=5 width=130) (actual time=0.001..0.001 rows=1 loops=11)                   |
|                      Filter: (NOT "FOUND")                                                                                                                          |
|                      Rows Removed by Filter: 0                                                                                                                      |
|                ->  Bitmap Heap Scan on "COMPONENT" "C_2"  (cost=29.59..31.85 rows=2 width=661) (actual time=9.744..9.856 rows=1 loops=10)                           |
|                      Recheck Cond: (("PREVIOUS"."PROJECT_ID" = "PROJECT_ID") AND ("DIRECT_DEPENDENCIES" ~~ (('%'::text || ("PREVIOUS"."UUID")::text) || '%'::text)))|
|                      Rows Removed by Index Recheck: 1                                                                                                               |
|                      Filter: ("ID" <> ALL ("PREVIOUS"."PATH"))                                                                                                      |
|                      Heap Blocks: exact=17                                                                                                                          |
|                      ->  BitmapAnd  (cost=29.59..29.59 rows=2 width=0) (actual time=9.573..9.573 rows=0 loops=10)                                                   |
|                            ->  Bitmap Index Scan on "COMPONENT_PROJECT_ID_IDX"  (cost=0.00..4.37 rows=379 width=0) (actual time=0.251..0.251 rows=9056 loops=10)    |
|                                  Index Cond: ("PROJECT_ID" = "PREVIOUS"."PROJECT_ID")                                                                               |
|                            ->  Bitmap Index Scan on idx_gin  (cost=0.00..24.97 rows=3409 width=0) (actual time=9.285..9.285 rows=2 loops=10)                        |
|                                  Index Cond: ("DIRECT_DEPENDENCIES" ~~ (('%'::text || ("PREVIOUS"."UUID")::text) || '%'::text))                                     |
|                SubPlan 7                                                                                                                                            |
|                  ->  CTE Scan on "CTE_MATCHES" "CTE_MATCHES_2"  (cost=0.00..0.02 rows=1 width=8) (actual time=0.001..0.001 rows=1 loops=1)                          |
|Planning Time: 0.874 ms                                                                                                                                              |
|Execution Time: 111.331 ms                                                                                                                                           |
+---------------------------------------------------------------------------------------------------------------------------------------------------------------------+
</pre>
</details>

For testing negative matches, I kept `is-ci@2.0.0` as leaf node. I misspelled the root node's name to emulate a no-match.

<details>
<summary>Old (113.563 ms)</summary>
<pre>
+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|QUERY PLAN                                                                                                                                                                               |
+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|Aggregate  (cost=38180.43..38180.44 rows=1 width=1) (actual time=113.510..113.512 rows=1 loops=1)                                                                                        |
|  CTE CTE_DEPENDENCIES                                                                                                                                                                   |
|    ->  Recursive Union  (cost=231.53..38143.71 rows=1632 width=78) (actual time=10.409..113.490 rows=11 loops=1)                                                                        |
|          ->  Bitmap Heap Scan on "COMPONENT" "C"  (cost=231.53..262.65 rows=12 width=78) (actual time=10.408..10.781 rows=1 loops=1)                                                    |
|                Recheck Cond: ("DIRECT_DEPENDENCIES" ~~ '%89dc699b-1518-49f9-ba74-04a8d0234400%'::text)                                                                                  |
|                Rows Removed by Index Recheck: 1                                                                                                                                         |
|                Heap Blocks: exact=2                                                                                                                                                     |
|                ->  Bitmap Index Scan on idx_gin  (cost=0.00..231.53 rows=28 width=0) (actual time=10.397..10.397 rows=2 loops=1)                                                        |
|                      Index Cond: ("DIRECT_DEPENDENCIES" ~~ '%89dc699b-1518-49f9-ba74-04a8d0234400%'::text)                                                                              |
|          ->  Nested Loop  (cost=29.28..3786.47 rows=162 width=78) (actual time=9.227..9.335 rows=1 loops=11)                                                                            |
|                ->  WorkTable Scan on "CTE_DEPENDENCIES" "CTE_DEPENDENCIES_1"  (cost=0.00..2.40 rows=120 width=131) (actual time=0.000..0.000 rows=1 loops=11)                           |
|                ->  Bitmap Heap Scan on "COMPONENT" "C_1"  (cost=29.28..31.52 rows=1 width=674) (actual time=9.211..9.319 rows=1 loops=11)                                               |
|                      Recheck Cond: (("CTE_DEPENDENCIES_1"."PROJECT_ID" = "PROJECT_ID") AND ("DIRECT_DEPENDENCIES" ~~ (('%'::text || ("CTE_DEPENDENCIES_1"."UUID")::text) || '%'::text)))|
|                      Rows Removed by Index Recheck: 1                                                                                                                                   |
|                      Heap Blocks: exact=18                                                                                                                                              |
…